# openapi.yaml
openapi: 3.1.0
info:
  title: WhatAgent API
  description: |
    The easiest way to send WhatsApp messages from code.

    ## Authentication

    All endpoints (except `/v1/status`, `/v1/webhooks/meta`, and `/v1/bootstrap`) require:

    ```
    Authorization: Bearer wha_<your-api-key>
    ```

    **Critical:** Use `Authorization: Bearer`, NOT `X-API-Key`. API keys always start with `wha_`.

    ## Quick Start

    1. **Bootstrap** — `POST /v1/bootstrap` with your BOOTSTRAP_SECRET to get a `wha_` API key
    2. **Send** — `POST /v1/messages` with `Authorization: Bearer wha_xxx`

  version: 1.0.0
  contact:
    url: https://github.com/rushdmowjee/whatagent
servers:
  - url: https://api.whatagent.dev
    description: Production
paths:
...
  /v1/auth/register:
    post:
      summary: Register — get a wha_ API key instantly
      description: |
        Zero-touch account creation. Provide an email and app name to get a `wha_` API key
        with 2,000 free messages (hobby plan). No Meta credentials required at this step.

        Idempotent: re-registering with the same email revokes the old key and issues a fresh one.
        Rate limited to 3 registrations per IP per hour.
      operationId: register
      tags: [Setup]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, app_name]
              properties:
                email:
                  type: string
                  format: email
                  example: you@example.com
                app_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: My App
      responses:
        '201':
          description: Registration successful — store the api_key, it will not be shown again
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                    example: wha_xxx
                  account_id:
                    type: string
                    example: acc_xxx
                  free_tier_remaining:
                    type: integer
                    example: 2000
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Rate limit exceeded — too many registration attempts from this IP
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key in `Authorization: Bearer wha_xxx` format.
        Keys always start with `wha_`.
        DO NOT use `X-API-Key` — WhatAgent uses standard HTTP Bearer auth.
    bootstrapSecret:
      type: http
      scheme: bearer
      description: |
        BOOTSTRAP_SECRET environment variable — used only for first-time setup.
        NOT the same as a wha_ API key.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...